Split test_reprs.py into respective files
